# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Battleship API",
    "description": "REST API for the Battleship game backend (session, preparation, gameplay)",
    "version": "2.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
...
  "paths": {
...
    "/api/v2/game/sessions/{sessionId}/players/{playerId}/field/shot": {
      "post": {
        "tags": [
          "Gameplay"
        ],
        "summary": "Make a shot",
        "description": "Resolves a shot at the given coordinate on the acting player's opponent field. Fails if the player id or coordinate is malformed, the session is unknown, it isn't this player's turn, the cell was already shot, or the session isn't in the IN_GAME stage.",
        "operationId": "makeShotByField",
        "parameters": [
          {
            "name": "sessionId",
            "in": "path",
            "description": "Session identifier",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "abc123"
          },
          {
            "name": "playerId",
            "in": "path",
            "description": "Shooting player's identifier",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "p1"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ParamCoordinateDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Shot resolved",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseShotResultDto"
                },
                "examples": {
                  "success": {
                    "description": "success",
                    "value": {
                      "shotResult": "HIT"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid input, unknown session, not this player's turn, cell already shot, or wrong stage",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ExceptionDto"
                },
                "examples": {
                  "playerIdInvalid": {
                    "description": "playerIdInvalid",
                    "value": {
                      "status": 400,
                      "errorMessage": "...",
                      "errorCode": "PLAYER_ID_INVALID"
                    }
                  },
                  "coordinateInvalid": {
                    "description": "coordinateInvalid",
                    "value": {
                      "status": 400,
                      "errorMessage": "...",
                      "errorCode": "COORDINATE_INVALID"
                    }
                  },
                  "sessionNotFound": {
                    "description": "sessionNotFound",
                    "value": {
                      "status": 400,
                      "errorMessage": "...",
                      "errorCode": "SESSION_NOT_FOUND"
                    }
                  },
                  "playerNotActive": {
                    "description": "playerNotActive",
                    "value": {
                      "status": 400,
                      "errorMessage": "...",
                      "errorCode": "PLAYER_NOT_ACTIVE"
                    }
                  },
                  "cellAlreadyShot": {
                    "description": "cellAlreadyShot",
                    "value": {
                      "status": 400,
                      "errorMessage": "...",
                      "errorCode": "CELL_ALREADY_SHOT"
                    }
                  },
                  "stageInvalid": {
                    "description": "stageInvalid",
                    "value": {
                      "status": 400,
                      "errorMessage": "...",
                      "errorCode": "STAGE_INVALID"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Unexpected internal error",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ExceptionDto"
                },
                "examples": {
                  "internal": {
                    "description": "internal",
                    "value": {
                      "status": 500,
                      "errorMessage": "...",
                      "errorCode": "INTERNAL"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v2/game/sessions/{sessionId}/players/{playerId}/state": {
      "get": {
        "tags": [
          "Gameplay"
        ],
        "summary": "Get gameplay state for a player",
        "description": "Returns the requesting player's and opponent's boards, alive-cell/ship counts, and win state. Fails if the player id is malformed, the session is unknown, or the player/opponent can't be resolved in this session.",
        "operationId": "getGameStateForPlayer",
        "parameters": [
          {
            "name": "sessionId",
            "in": "path",
            "description": "Session identifier",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "abc123"
          },
          {
            "name": "playerId",
            "in": "path",
            "description": "Requesting player's identifier",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "p1"
          }
        ],
        "responses": {
          "200": {
            "description": "Gameplay state for the requesting player",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseGameplayStateDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid player id, unknown session, or player/opponent not found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ExceptionDto"
                },
                "examples": {
                  "playerIdInvalid": {
                    "description": "playerIdInvalid",
                    "value": {
                      "status": 400,
                      "errorMessage": "...",
                      "errorCode": "PLAYER_ID_INVALID"
                    }
                  },
                  "sessionNotFound": {
                    "description": "sessionNotFound",
                    "value": {
                      "status": 400,
                      "errorMessage": "...",
                      "errorCode": "SESSION_NOT_FOUND"
                    }
                  },
                  "playerNotFound": {
                    "description": "playerNotFound",
                    "value": {
                      "status": 400,
                      "errorMessage": "...",
                      "errorCode": "PLAYER_NOT_FOUND"
                    }
                  },
                  "opponentNotFound": {
                    "description": "opponentNotFound",
                    "value": {
                      "status": 400,
                      "errorMessage": "...",
                      "errorCode": "OPPONENT_NOT_FOUND"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ExceptionDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "integer",
            "format": "int32",
            "description": "HTTP status code",
            "example": 400
          },
          "errorMessage": {
            "type": "string",
            "description": "Human-readable error message"
          },
          "errorCode": {
            "type": "string",
            "description": "Stable, machine-readable error code identifying the exception type",
            "example": "SESSION_NOT_FOUND"
          }
        }
      },
...
      "ParamCoordinateDto": {
        "type": "object",
        "properties": {
          "row": {
            "type": "integer",
            "format": "int32",
            "description": "Zero-based board row"
          },
          "col": {
            "type": "integer",
            "format": "int32",
            "description": "Zero-based board column"
          }
        }
      },
      "ResponseShotResultDto": {
        "type": "object",
        "properties": {
          "shotResult": {
            "type": "string",
            "description": "Outcome of the shot",
            "enum": [
              "MISS",
              "HIT",
              "DESTROYED"
            ]
          }
        }
      },
...
      "CellDto": {
        "type": "object",
        "properties": {
          "row": {
            "type": "integer",
            "format": "int32",
            "description": "Zero-based board row"
          },
          "col": {
            "type": "integer",
            "format": "int32",
            "description": "Zero-based board column"
          },
          "ship": {
            "$ref": "#/components/schemas/Ship",
            "description": "Ship occupying this cell, if any (engine record: shipId, shipType, shipDirection, shipSize)"
          },
          "hasShot": {
            "type": "boolean",
            "description": "Whether this cell has already been shot at"
          },
          "isAvailable": {
            "type": "boolean",
            "description": "Whether this cell is available for placing a ship"
          }
        }
      },
      "ResponseGameplayStateDto": {
        "type": "object",
        "properties": {
          "playerName": {
            "type": "string",
            "description": "Requesting player's display name"
          },
          "playerNumberOfAliveCells": {
            "type": "integer",
            "format": "int32",
            "description": "Number of the player's board cells still occupied by unsunk ships"
          },
          "playerNumberOfAliveShips": {
            "type": "integer",
            "format": "int32",
            "description": "Number of the player's ships not yet fully sunk"
          },
          "playerField": {
            "type": "array",
            "description": "Player's own board, as a 2D array of cells",
            "items": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/CellDto"
              }
            }
          },
          "opponentName": {
            "type": "string",
            "description": "Opponent's display name"
          },
          "opponentNumberOfAliveCells": {
            "type": "integer",
            "format": "int32",
            "description": "Number of the opponent's board cells still occupied by unsunk ships"
          },
          "opponentNumberOfAliveShips": {
            "type": "integer",
            "format": "int32",
            "description": "Number of the opponent's ships not yet fully sunk"
          },
          "opponentField": {
            "type": "array",
            "description": "Opponent's board as visible to the requesting player, as a 2D array of cells",
            "items": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/CellDto"
              }
            }
          },
          "hasWinner": {
            "type": "boolean",
            "description": "Whether the game has a winner"
          },
          "winnerPlayerName": {
            "type": "string",
            "description": "Winning player's display name, if any"
          },
          "isPlayerActive": {
            "type": "boolean",
            "description": "Whether it is currently this player's turn"
          },
          "isPlayerWinner": {
            "type": "boolean",
            "description": "Whether this player has won the game"
          },
          "isOpponentReady": {
            "type": "boolean",
            "description": "Whether the opponent has readied up"
          }
        }
      },
      "Ship": {
        "type": "object",
        "properties": {
          "shipId": {
            "type": "string"
          },
          "shipType": {
            "type": "string",
            "enum": [
              "PATROL_BOAT",
              "SUBMARINE",
              "DESTROYER",
              "BATTLESHIP",
              "CARRIER"
            ]
          },
          "shipDirection": {
            "type": "string",
            "enum": [
              "HORIZONTAL",
              "VERTICAL"
            ]
          },
          "shipSize": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
    }
  }
}